Stable
Improved user feedback during the initial stage:

Added status updates for the number of processed images.
Displayed CPU and memory utilization information.
Added a preview image during processing:

Implemented the 'update_preview_image' function to display a thumbnail of the current averaged image.
Updated the UI to include a preview image label.
Added code to update the preview image during the processing loop.
Implemented a "Stop" button to cancel the process:

Added a global variable 'stop_process' to monitor if the process has been stopped.
Created a 'stop_process' function to set the 'stop_process' variable to True when the button is clicked.
Modified the processing loop to break if the 'stop_process' variable is set to True.
Updated the UI to include the "Stop" button.
Added a "finished" message and played a beep sound when the processing is completed:

Modified the 'update_ui' function to set the status label to "Finished!" and play a beep sound when the "done" message is received.
Fixed an issue with the missing 'update_preview_image' function:

Added the 'update_preview_image' function to update and display the preview image in the UI.
Imported the necessary 'ImageTk' module from PIL.